Fix #1004 - Create separate model objects for arrival-and-departure, trip-plan, and destination-reminder data (#1005)
